requirements.txt
asn1crypto==0.24.0
astroid==1.6.3
beautifulsoup4==4.6.0
bleach==2.1.3
boto3==1.7.80
botocore==1.10.80
cachetools==2.1.0
Cartridge==0.12.0
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
colorama==0.3.9
comtypes==1.1.4
cryptography==2.6.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-contrib-comments==1.8.0
docutils==0.14
filebrowser-safe==0.4.7
future==0.16.0
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-dataproc==0.2.0
google-cloud-logging==1.6.0
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grappelli-safe==0.4.7
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.11.3
idna==2.6
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
Mezzanine==4.2.3
mrjob==0.6.4
mysql-connector-python==8.0.15
oauthlib==2.0.7
Pillow==5.1.0
protobuf==3.6.1
psycopg2==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pygame==1.9.4
pylint==1.8.4
PyMySQL==0.9.3
PyPDF2==1.26.0
pystache==0.5.4
python-dateutil==2.7.3
python-dotenv==0.10.1
python-telegram-bot==11.1.0
pytz==2018.3
PyYAML==3.13
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
sqlparse==0.3.0
static==1.1.1
static3==0.7.0
tzlocal==1.5.1
urllib3==1.22
virtualenv==15.2.0
webencodings==0.5.1
wrapt==1.10.11
xhtml2pdf==0.2.1
